+++ release_notes.md
@@ -1,4 +1,14 @@
 - - -
+## [1.1.0-mailbox-subscriptions](https://github.com/argon/push_notify/releases/tag/1.1.0-mailbox-subscriptions)
+Date: 25 September 2016
+
+### Features
+
+- Mailbox subscriptions
+
+### Fixes
+
+- - -
 ## [1.0.0-initial-release](https://github.com/argon/push_notify/releases/tag/1.0.0-initial-release)
 Date: 25 September 2016